+++ log/level.go
@@ -1,9 +1,5 @@
 package log
 
-import (
-	log "github.com/sirupsen/logrus"
-)
-
 type LogLevel int
 
 const (
@@ -13,11 +9,3 @@ const (
 	WARNING
 	ERROR
 )
-
-var levelMapping = map[LogLevel]log.Level{
-	TRACE:   log.TraceLevel,
-	DEBUG:   log.DebugLevel,
-	INFO:    log.InfoLevel,
-	WARNING: log.WarnLevel,
-	ERROR:   log.ErrorLevel,
-}